requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
pypdf>=4.0.0
openai-whisper>=20231117
yt-dlp>=2024.1.0
feedparser>=6.0.0
yfinance>=0.2.40
APScheduler>=3.10.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
lxml>=5.0.0
